requirements.txt
chromadb
colorama
pyyaml
requests
# Optional provider:
# ollama
